Finish Ex25 eventcapture bubbling propagation once
events happen on all overlapping elements all the way up to your os

browser does a capture that starts from outer all the way to inner and
fills a stack. the elements in the stack pop off in a. this is called
bubbling.

capture is set to false by default which causes the events to run on
bubble up (pops events from the top of capture stack). if set to true
events will run on bubble down (deques events from bubble que).

(e.stopPropogation() stops bubbling from happening)

once de-binds an event listener from the element it listens on so that
event does not trigger anymore. useful for shopping buttons when you
only want one click to work/ add an element to cart.